refactor: minor refactor of the `IntegrationManager`
